automation

+++ make-docs.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 import os
 import shutil
-import subprocess
 
-subprocess.run(['mvn', 'clean', 'javadoc:javadoc'], check=True)
+if os.system('mvn clean javadoc:javadoc') != 0:
+    exit(1)
 
 if os.path.exists('docs'):
     shutil.rmtree('docs')
